refactor: reorganize arXiv and citation tools into dedicated directories
Moved arXiv tools to dedicated arxiv/ directory:
- ArxivDownloadTool, ArxivMetadataTool, ArxivSearchTool
- Better separation from LaTeX-specific tools

Renamed citations/ to citation/ for consistency:
- CrossrefDoiTool, CrossrefSearchTool
- constants.ts, rateLimiter.ts

Updated all imports and registry references accordingly.
