+++ src/tools/citation/CrossrefSearchTool.ts
