+++ src/tools/index.ts
@@ -16,6 +16,6 @@ export * from './registry';
 export * from './wolfram';
 export * from './texcount';
 export * from './latex';
-export * from './citations';
+export * from './citation';
 export * from './web/WebFetchTool';
 export * from './web/WebSearchTool';
