+++ src/tools/latex/index.ts
@@ -1,5 +1,2 @@
-export { ArxivDownloadTool } from './ArxivDownloadTool';
-export { ArxivMetadataTool } from './ArxivMetadataTool';
-export { ArxivSearchTool } from './ArxivSearchTool';
 export { ExtractLatexFiguresTool } from './ExtractFiguresTool';
 export { ExtractTikzFiguresTool } from './ExtractTikzFiguresTool';
